requirements.txt

# Minimum Django and REST framework version
Django<2.0
djangorestframework>=2.4.3
django-friendship>=1.5.0

# Test requirements
pytest-django
pytest
pytest-cov
factory-boy
flake8

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1